feat:锁定库存重构

+++ adapter/src/main/java/com/lzb/adapter/rpc/inverntory/dto/LockStockReqDetail.java
@@ -13,7 +13,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class LockStockReqDetail {
 
-    private final long skuId;
+    private final int skuId;
     private final int num;
 
 }
